Removed old (pre-Ent) vegetation model from the code. It has not been used in any of the production rundecks for quite a few years and it was not maintained anyway. No USE_ENT cpp flags are needed any more, since Ent is now used anyway.
If for some reason somebody needs to use old (FK) photosyntesis
scheme in the model, please let me know and we will enable this
functionality through Ent interface (the code is there, it was just
not mointained for a while).

A few notes on ths change in the code:

1. Weird code in RAD_DRV.f which used #ifndef USE_ENT was replaced with

 #ifdef RESET_SURFACE_FRACTIONS_ON_SOUTH_POLE
        call get_param('vdata',vdata(1,1,:),size(vdata,3),
     &      default=(/0d0,0d0,1d0,0d0,0d0,0d0,0d0,0d0,0d0,0d0,0d0,0d0/))
 #endif

If you use this code please check.

2. Sorry, have replaced single "'" with "''" in a few comments to allow
parsers to work properly.

3. In flammability_drv.f "#ifdef USE_ENT" is all over it. So, I left
it to the owner to decide what to do with it. Just added
"#define USE_ENT" at the top of the file.