Major structural update in chemistry, to allow for automated definition of reaction rate indices. Not all reactions are implemented yet, but the concept is ready and will be implemented in small pieces, since no functionality breaks and the results are not changing. I will make it clear when the last commit is pushed, which will allow the addition of reactions rearrangement of the existing ones to be transparent.
This update (and the ones that will follow in the future, until the full
implementation) requires line rearrangements in the code, so diff will show
more lines to be different from what they actually are, due to several changes
in line continuations.

The same approach will be followed for the photolysis rates, also in the future.