This is an extension to get_input_data which allows the downoald of annual data files located in a subdirectory. Sorry, current implemetation is rather messy and has limited functionality, but it allows those who need this functionality to proceed. I may clean it uplater.
As it is now, the the input data can be stored as one file per year
in the following format:
  prod_input_files/
                  <dataset_name>/
                                <year1>.nc
                                <year2>.nc
                                 .....
                                <yearn>.nc

<yeark> should be in a 4-digit format, i.e. the file names
will be : 1850.nc, 1851.nc etc.

When used locally, get_input_data copies all files from such
subdirectories. WHen used to download data from a remote repository
(i.e. dataportal), it requires an additional option:

 -y <year_start>:<year_end>

When present, the datafiles from <year_start> to <year_end> will be
downloaded. For example,

 ../exec/get_input_data -w -y 1972:1980 Efoobar

downoalds all the data needed to run the simulations from 1972
until 1980. If -y is not present, a single year 1850 is used
by default (preindustrial conditions).

The rundeck should specify the path to the subdirectory with
the dataset, i.e.

FOO=<dataset_name>

No trailing slash or other special marks are needed. The script
will identify automatically that the entry is a directory.
(It uses a file with the list of subdirectories for this purpose).