Slightly improved version of get_input_data. The script now keeps the table of contents for the entire directory structure. For each file with annual data it checks first if the file is present in "dirlist" and if not, skips the attempt to download.
This allows the directories with annual data to be "sparce",
i.e. to contain the data only for some years (for example, one
file per decade). The script (if executed with -y year_start:year_end)
will still check the table of contents for the presence of all
annual files between year_start and year_end, but will attempt
to download only the ones which are present. A warning will be
printed for the files which don't exist.