Removed spurious CONST.f files, as per discussion with Igor Aleinov and Kostas Tsigaridis:
Aleinov:
Old files pop up from time to time after messy merges or cherry picks.
Unfortunately, if this is not noticed immediately, it is not obvious to decide
if some file is not needed, especially if these are system files not owned
by anybody in particular. If I remember correctly, CONST.f  was renamed by
Tom Clune, so the original file should be deleted.

Tsigaridis:
Your detective work is right. What you hadn't discovered though is that my restoration of the file was accidental, and, when discovered, nobody took the courage to delete CONST.f once more. As far as I understand, you can safely delete it, since no rundecks (should) use it.