Please read if you use the tajl or taij files!
A lot of atmospheric tracers diagnostics have been renamed in the tajl and taij
files, in an effort to automate diagnostics definition. The following
assumptions are now followed for these two files:

If a tracer is present in the aqchem_list list, it will get a diagnostic named
TRACERNAME_aqchem in taij, and TRACERNAME_mc_cloud_aqchem and
TRACERNAME_ss_cloud_aqchem in tajl.

If a tracer has surface sources, as defined by ntsurfsrc, it will get a
diagnostic named TRACERNAME_SOURCENAME_src, in both files.

If a tracer has aircraft emissions, as defined by do_aircraft (which is
defined by the presence of a TRACERNAME_AIRC entry in the rundeck), it will get
a diagnostic named TRACERNAME_aircraft_src, in both files.

If a tracer has biomass burning emissions, as defined by nBBsources (defined
in the rundecks per tracer), it will get a diagnostic named
TRACERNAME_biomass_src, in both files.

Volcanic sources are still not automated, but their name became consistent in
both files, and it is TRACERNAME_volcanic_src.

All above-mentioned diagnostics now have CF-compliant units (kg m-2 s-1 in
taij and kg s-1 in tajl).

All OMA, MATRIX, and TOMAS still maintain the assumption that if a SO2 source
is defined, it is automatically used for SO4 as well, with a 2.5% yield.

MATRIX has some additional assumptions for all tracers with offline sources,
both in number and mass.

***************
The above-mentioned automations work in diagnostics definitions only. Under
many circumstances they will also work in diagnostic accumulation, but this
needs to be tested by the individual users that add such enhancements in the
code
***************