Added more options for dust emission, which can be controlled with rundeck parameter imDust. The available options are now as follows:
0: scheme using PDF of wind speed (current default).
1: prescribed AeroCom emissions.
2: legacy emission scheme using third power of wind speeds (only works
   with 72x46 horizontal resolution).
3: use size distribution of AeroCom for model calculated dust/mineral
   emission flux (PDF scheme).
4: use the size distribution of the mineral version of the dust
   aerosols for the emitted dust flux (PDF scheme).
5: same as 4, but no dust emission where no AeroCom sources

The results have not changed for imDust=0, the current default, with
this commit. The idea is to transition to imDust=4 as the new default,
though, which is based on recent literature and makes the one-type
dust aerosols more an idealization of the dust aerosol version with
different minerals type.

The different imDust options work for all aerosol physics versions
(OMA, MATRIX, TOMAS) with dust emission.

The commit comes with some modifications in the template includes for
dust related input files and parameters in the templates
sub-directory.

An updated version of the input file with the mineral fractions at
emission is used. It is nearly the same as the previous one, but with
some small artefects reduced in the data, which were caused by
interpolation of discrete soil texture types on a high resolution to a
lower resolution. The file is also needed now for the templates with
one-type dust for imDust=4 and 5.

A new rundeck parameter, scaleDustEmission, has been introduced, which
scales the total dust emission flux. It is supposed to be used for
options imDust=1 and 3-5, while fracClayPDFscheme and
fracSiltPDFscheme are supposed to be set to 1.0 for these imDust
options. For imDust=0, fracClayPDFscheme and fracSiltPDFscheme are
still supposed to be used as is, and scaleDustEmission=1.0, although
latter parameter could be used to additionally scale the totally
emitted dust mass up or down.

The value for scaleDustEmission currently used in the templates is
preliminary. The calibration still has to be properly done for the
different imDust options, separately for OMA, MATRIX, and TOMAS, and
then again for the different aerosol physics versions with nudged
winds.

This commit also introduces two more rundeck parameters:

vegetationERS. If set to 0, it switches off the vegetation proxy data
filter (and the reading in of the according input file), which is
based on European Remote Sensing (ERS) satellite retrieved roughness
lengths and which is used to control the magnitude of the dust
emission, depending on the geographic location. The default is 1.

prefDustSources. If set to 0, it switches off the preferred dust
sources filter (and the reading in of the according input file), which
also depends on the geographic location. The default is 1.