Submit TOMAS rundecks with CMIP6 emissions files/settings. As usual with TOMAS, this doesn't indicate an endorsement  that it's working - just keeping it alive. But it compiles, and decks will be used for standard regression tests soon.
Increase ktajls for TOMAS compilations (needed when 
SHINDELL_STRAT_EXTRA and ACCMIP_LIKE_DIAGS
directives are on.)

However, I am turning those two directives off, for the moment,
in E6TtomasF40.R and E6TmatrixF40.R, as they mess up 
scaleacc because of the netCDF variable number limit. Once that
is dealt with, I recommend turning these back on in case needed
for MIP-submissions. Leaving on in the equivalent OMA decks.