removing FLTRAN file from rundecks.
First a note that since Tiehan's push 3 days ago (RADN9 file
reading is netCDF) and Kostas's 1 day ago (to allow template decks
to run by listing that nc file) the photolysis would still be
"seeing" the older solar flux we've been running with the last year
or so. (And therefore would have been out of sync with the
rad code since those commits).

This change (and code changes pushed with this) tells the decks to
no longer read from the ascii FLTRAN file but directly from the
new RADN9 file. This puts us back in sync. But some notes:

(1) The logstanding use of rad_FL=1 in the parameters is still valid.

(2) This push therefore changes the chemistry simulation to see the
CMIP6 solar input. As the new solar file reading just arrived, we do
not yet know how the chemistry looks. I will run a longer run and
run results by Drew Shindell. Expecting modest changes.

(3) See other commit notes in this combined push regaring backwards
compatiblity with the ascii FLTRAN input method. I note here that
the answer is not exactly the same as with the FLTRAN ascii input
because that ascii file had only a few significant figures whereas
Tiehan's netCDF input has the normal amount of single-precision
netCDF significant figures.