A few changes to the scripts necessary to run the model on Pleiades. This includes the functionality some of you were requesting - the model now remebers the modules it was compiled with (was necessary, because compute nodes on pleiaes don't inherit the environment from the login nodes).
The names of modules are stored in "modules" file in the run directory.
The runE script gets them from there. Reto's "run" script is not aware
of this functionality yet.

I don't see any reson why one would not want the scripts to work this
way, but let me know if you need a switch to disable this functionality.
As a hack, one can always delete "modules" file, in which case
the modules will not be reloaded.