Manual cherry-pick of 2011 AR5-branch commit 50d19618c8631cf16b608204254b114e7f68e8ec. No effect on results unless opting in.
Log message from the 2011 commit:

To help LGM runs cope with extremely thick Arctic sea ice,
the latlon ice advection now has a variation of the coding
added to the cubed-sphere configuration to prevent unrealistically
large coastal maxima of ice thickness (CVS rev. 1.8 of ADVSIcs.f).
To activate the new coding, EXPEL_COASTAL_ICEXS must be CPP-defined.
Further adjustments may be necessary.

A corrective transport velocity is added to the velocities predicted
by the ice dynamics.  It is nonzero only between gridcell pairs
for which at least one cell is incompletely "ocean connected".
(Connectedness is defined using the number of adjacent cells
having a nonzero ocean fraction).  The corrective velocity
along a particular axis is proportional to the gradient of ice
thickness, and if two cells have a different connectedness along
that axis, the velocity is only nonzero if the ice is thinner
in the cell having higher connectedness.