Removed USE_ENT option from all templates, since Ent is now a default vegetation scheme in the model.
Also removed the remaining instances of USE_ENT from all scripts,
documentation, and comments in the code.

The only source file where this option remains (set as default)
is model/flammability_drv.f . It is used extensively over the code,
so I am leaving it to the owner of this code to do the fixes.