Adding a rundeck parameter "aerosols_affect_photolysis" that controls whether the aerosol profile should be taken into account in fastj or not. This only impacts OMA, since in MATRIX and TOMAS this effect is not taken into account.
I am not adding it in the template rundecks because very few instances
will be using it. To disable aerosols in fastj, set

aerosols_affect_photolysis=0

in the parameters section of the rundeck. Note that water and ice clouds
will still be impacting photolysis rates, and this applies to MATRIX and
TOMAS as well.