Change some sections of optional chem debugging print statements to survive traps. (In the cases where these were format statement changes only: they might not be perfect, but·they work now.)
One case was potentially important: The masterchem call to
printSS27x2Etc debug diag writing was inside an L-loop and
yet the routine has its own L-loop. Hence having that on
would mess up the actual chemistry calculations because L would be
41 upon return for a 40-layer model. (Only relevant if you had
made a special copy of the MOLEC file to set prnchg set to t).
I have moved the call outside of the L loop and for good measure·
changed the interior loop to be over LPRINT instead.