avoid is_fbsa() test in flammability code.
Under gfortran only, a check that a certain file used for
the fire model is fbsa fails. Since all this did was stop
the model and that's what should happen if you use the
wrong file anyway, I am simply commenting out the test as
guidance for future programmer to swap to netCDF reading.

Hoping this allows the template with fire model engaged to
survive regression tests under gfortran.

(The fbsa file in question, btw, is not a pure "GISS binary").