Initial stab at maintaining a 102L OMA template.
It is provided for model development purposes; not tuned or approved
for Science.

A substantial increase in MAX_RPARAMS was required to get this to
compile, and I think we're not supposed to use CPP logic to do that
(e.g. increase only if tracers are on) so I simply increased it.

Right now this was a simple merge of the F40 OMA template and the
LLF102.R, except for:
chose not to 'include' condensationF96_params, as these params are
    usually tuned separately for this configuration anyway
setting the chemistry tuning parameters to what we had the last time
    we ran a 102L with chemistry (essentially all off; this is
    different from the 40L settings). See chemistry_params_102_CMIP6
    vs. chemistry_params_CMIP6