Eliminating the MOLEC input file.
The file had three purposes:

1) Define some diagnostic flags. These are now controlled by rundeck options,
   commented-out by default.
2) Define the transported tracers in chemistry. This is now unified with the
   rest of the tracer definitions.
3) Define the non-transported tracers in chemistry, as well as the well-mixed
   gases that are relevant for chemistry, e.g. O2, H2, M. These are now defined
   in a similar way as the transported tracers.

One important change for people familiar with the chemistry internals is that
the variable ay (which was storing the tracer names read in from MOLEC) is now
called trchemname. Another important change is that the variables iprn,jprn,lprn
are now a 3-element array: ijlprn.