more megan integration to get a source to show up
(an Isoprene source show up now in the setup hour, though
no evaluation of its size and several workarounds are still
in place)

* do_megan(n) changed from logical to integer so it can be used to
remember the index of megan source in the 2D surface source array.
(This is set when diags are set up after addSurfaceSource call in
initTracerMetadata.f. In there, moving call to be before check for
ntsurfsrcmax violation after fire sources are accounted for, so that
total will include megan source in that check).

* do_megan is passed to readSurfaceSources so it can skip reading
for that index. Might be ugly, but I deemed it less ugly than
determining how many files to read, e.g. for a tracer that might
have megan and 3D biomass burning, etc. etc.

* move call to biogenicEmissions_drv to before the tracers loop so
that sfc_src can be used to fill in trsource within the loop. Exempt
megan source from the COSZ1-based sunlight weighting other Isoprene
sources have (since it does its own).

* some changes to megan.F90 comments