Corrected 2 serious bugs in ParameterizedOrbit.
The first bug was that the time at periapsis must be recomputed any
time that the longitude at periapsis changes.  The impact is a large
shift in the time of year for the seasons.

The second bug was that the duration of the mean day was not being set
in this case.  It ended up with a default value of 0.  The only known
impact was that RAD_DRV would compute zenith angles based upon the
time at the beginning of a day rather than the time at the _middle_ of
the day.  Possibly the code should be changed to instead derive the
mean time of day when the orbital and rotatioanl periods are set.