This is a proper fix for the particle tracer concentration problem in irrigation water. The original implementation had a positive feedback due to the way the tracer concentration was assigned for "fossil" water. The new implementation assumes that "fossil" water has no particle tracers.
One should mention, that this fix was applied only to particle tracers.
Similar problem can happen to water tracers proper, if there is
a fractionation during the evaporation (as in the case of water isotopes).
Somebody has to make a decision on which should be the concentration
of such tracers in "fossil" water (it can't be gtracer or concentration
in the lake water, which would cause a positive feedback).