Changes to reading of continuous volcano source:
(1)  PSREF file converted to netcdf

(2)  The same SO2_VOLCANO file is now read for all horizontal grids.
     Previously, the reading in the case of a lat-lon grid assumed the
     input file was on the model grid.

These changes are not testable as of the repository state before this commit,
since there is a pre-existing problem running OMA.    However, the coding
was verified in a prior version of the repository.   Change (2) results
in very slight differences for 2x2.5 runs.