Reviving ability to manipulate subsets of tracers.
The core functionality of this commit is mostly work of Tom Clune,
while I only made sure that the code works for all three aerosol modules
used in the model. Tom's message, slightly modified to reflect my changes, is:

Added a subroutine in TRACER_COM::get_tracer_subset_indices() that
returns an allocatable array of integers that are the indices for tracers
that have a specified property.

This is demonstrated in TRACERS_AEROSOLS_Koch.f for BC species for OMA, AMP, TOMAS.