Modifications to tracer coding so that that tracer water tracks humidity.   This commit incorporates parts of AR5_v2_branch commit f2a613454dc2e4804f6518d10be8c63b3eaed566 whose log message was:
Corrections:

1. In the CTEI mixing to account for the fact that FSSL now varies with layer
(FSSL = 1-FMC1 below MC cloud top, FSSL=1 above).

2. To make two tracer calculations in the moist convection scheme
consistent with those for temperature and humidity.  The first
concerns the entrainment of tracer into updrafts, and the second
the incorporation of tracer mass into downdrafts at their formation
levels.  This correction is commented out pending evaluation of
its effects on CADI runs; it is necessary for agreement
between humidity and the "Water" tracer.  The recent activation of
precip re-evaporation above MC cloud base seems to be the
mechanism by which this inconsistency started to generate
disagreement between humidity and "Water".  QM1, which defines
the "environment" for subsequent plumes, is set before the
re-evaporation of precip from a given plume, and can therefore
differ from QM above cloud base.