I added function wv_psat to Utilities.F90 for calculating the saturation water vapor pressure, using the same approximation that is used in the core of function qsat for calculating the saturation specific humidity. wv_psat can be accessed from anywhere in the model. Also, in TOMAS, I replaced the call to a TOMAS specific function (VPRES) for calculating the saturation water vapor pressure with a call to wv_psat. The TOMAS specific function uses a polynomial approximation that becomes negative for temperatures around 210 K, crashing the model in such instances.
From the point of view of model philosophy, it is more sound to
consistently use the same calculations for basic physical processes
everywhere in the model, anyway. Differences in the results between
the different aerosol physics versions should come from actual
differences in the aerosol physics, not from differences in
calculations of basic physical processes.

The commit does not change any results, except for TOMAS.