Another CO2-scaling change affecting the 102-layer model only; the previous commit did not work well and a new method was needed to deal with that case.
No changes for the other layerings (20, 40, 96 layers) and only the first 4
columns of CO2profile are useful, so we can go back to the previous table
(CO2profile.Jul2017.txt).

No CO2profile file is needed for the new 102-layer model treatment; it is
tuned to this particular layering, but should also work if the layering below
50 mb is modified (refinement or coarsening). If KPFCO2=3 or 4, a
latitudinally and seasonally dependent vertical CO2 scaling is applied, if KPFCO2=4 (the default for that layering), the up-flux corrections are modified
as well (part 2).

Andy Lacis notes that this method works better in combination with
NRAD=1 (calling the radiation every half hour rather than every 2.5 hours),
but that necessarily makes the model substantially slower. He also warns that
the part 2 might cause unrealistic temperatures and we might have to
apply it with a lower frequency.

A small correction in TAUGAS was made that only affects the "TAPER_UTCF" case.