Allow .gt. 1 MEGAN source for a tracer.
(Though allowed, model still only has one source per tracer until next
commit on this work).

The do_megan array [which used to point to the index of the MEGAN
source] is removed in favor of setting isMegan=.true. for a tracer
surface source object, meaning:
trc%surfaceSources(ns)%isMegan = .true.
Note this can be set using an optional argument to addSurfaceSource()
calls.

The sourceName (meaning trc%surfaceSources(ns)%sourceName) of the tracer
surface source object is used to match the source to a MEGAN sub-species.