Switch to netCDF reading for 5 Shindell IC files
These: N2O_IC, CFC_IC, CH4_IC, Ox_IC, CO_IC
replaced with netCDF versions. The number of layers in the input is
still a harcoded parameter, not read from the files. But I think this
will be easy to change now - each of these is now read / interpolated
in the same subroutine and the input pressure levels are already read
from the input files.

This commit will break 36x24 and C90 decks until files are made for
them in a later commit.