Short version: suppress AIJ outputs for levels above the model top.
Long version:

A recent commit extending the number of fixed-pressure levels for certain AIJ
diagnostics also caused a discrepancy in the number of such levels
between AIJ and AGC because the former was then unconditionally outputting
at the full set of possible levels no matter the location of the model top.

Under the assumption that this glitch was facilitated by confusion about
the actual set of available levels versus the maximum possible for
a top at infinity, I have unilaterally decided to swap the roles of
the misleadingly named KGZ and KGZ_max (according to git blame, the names were
chosen in 2001 by a person who is not here to protest).  Now KGZ_max is the
maximum possible, and KGZ the actual number of available pressure levels.