add cyclic to fire model + coupled_chem=0 reading
* subroutine aerosol_gas_chem_prep - allow overriding of the year for
reading from OFFLINE_CHEM, using aer_int_yr. When that is in use, set
cyclic=T. [Currently this should do nothing, as there are only 12
months in the input files, but this may help in the future].

* subroutine masterchem_prep - same thing for reading from OFFLINE_AERO,
but overriding uses o3_yr. In this case it might matter, as there are
currently 2 time slices in the sample input file.

* subroutine readFlamPopDens - allow cyclic to be true in the reading
of population for fire model, based on whether ABS(o3_yr) > 0. [These
are annual data though, not monthly]