Replace gsfMGOLjal hack for CH4 emissions
Applicable to defined WATER_MISC_GRND_CH4_SRC, for both Shindell and
Lerner tracers.

Formerly, as a way of minimizing special coding, we turned on Ocean,
Lake, and Misc. Ground source for CH4 emissions by listing a blank file
with an intentionally-obscure source name metadatum. When the model
detected this special source name, it filled in the source.

New solution: we had an isMegan logical as part of a tracer source
object, and this was used to skip reading of a source tagged with that.
I generalized this to be an integer called skipReason instead. There are
no longer any files read for ocean/lake/misc ground source - the
condition to fill in the source is now:
(pTracer%surfaceSources(ns)%skipReason==itsCH4MGOL)

As a consequence of that generalization, the new way to 'tag' a MEGAN
source is: skipReason=itsMegan (instead of isMegan=.true.)

The templates were changed to remove these CH4 source files, renumber
the others, and change the ns_wet parameter for interactive wetlands.

The change in results you will see for the E6TomaF40int template
is due to the reordring of its CH4 sources (and thus reordering of their
summation), not this new coding, per se. [E.g. no change in Lerner
template even though it uses WATER_MISC_GRND_CH4_SRC, because its source
happens to be in the same position when you delete the formerly-
associated file and use the new code.]