Chemistry retune and updated fastj solar input
In the templates: rad_input_files (and rad_input_files_noCO2profile
but not Emaobio_08112017.R, since that had a date in its name) I am
changing the RADN9 solar input file. The solar input to the radiation
code does not change -- only what is sent to the photolysis code in
use by Shindell chemistry. Reason: for E2.1 we had a miscommunication
and re-binned the solar input to fastj input in the 'old way' from the
coarser GCM solar input file. This inadvertantly delayed the
improvement (of higher-resolution rebinning from the CMIP6 solar input)
to E3. That improvement is contained in this push.

To go along with that change in fastj input (primarily, though some
smaller amount of retuning would have been necessary anyway), is a
retuning of the Shindell chemistry via rundeck parameters (changed
in chemistry_params, chemistry_params_CMIP6, and
chemistry_params_CMIP6_noBBURN).

Obviously, expect changes in answer for templates that compile the
chemistry.