remove chemistry ss array from restart files
also zeroing these photolysis rates each chemistry
step, as they only get filled in in sunlight.

No change in results except: (a) the ss array in the rsf files,
of course, and (b): ACCMIP diagnostic change for ijlt_pO1D,
but I think this is in fact a bug fix, because it seems
like that diag should be accumulating 0 in darkness, not
the production based on the last photolysis rate available.

With luck, further commits will replace ss(4D) with the zj(2D)
variable.