Several simplifications of Shindell coding
The codirect and GLT diagnostic tracers are now set to be active
for any run with Shindell chemistry. This allowed for the removal
of SHINDELL_STRAT_EXTRA directive & category of tracers.

ACCMIP_LIKE_DIAGS diagnostics are now always on for Shindell
compilations. Exceptions: for ijlt_prodSO4aq & ijlt_prodSO4gs diags,
it seemed better to tie those to:
{TRACERS_AEROSOLS_Koch or TRACERS_AMP or TRACERS_TOMAS}
instead of to Shindell. So ACCMIP_LIKE_DIAGS now removed.

Note that in RAD_DRV there is no longer a model stop if
rad_forc_lev > 0 and ACCMIP_LIKE_DIAGS (now TRACERS_SPECIAL_Shindell)
is on. This was there because ACCMIP required TOA aerosol radf diags.
It is now up to the user to have this set as desired (current E6 tracer
templates have this set to 0).

SHINDELL_STRAT_EXTRA and ACCMIP_LIKE_DIAGS removed from all templates:
The E6 oma/matrix/tomas templates had these on and shindell on, so
should be no effect. But many older templates (like E4*) had these off.
Therefore those configurations will have more tracers and diags now.

Removing the dependance of GLT diagnostic tracer on itime_tr0(). This
tracer has a linear-in-time increase in overwriting source. Previously,
itime_tr0 was employed as a way of smoothing the transition from a
spinup run to an experiment run - one could set itime_Tr0(n_GLT) in the
rundeck, and time was measured as:
(Itime-ItimeI-itime_tr0(n_GLT))
Now one can use a dedicated rundeck parameter: GLToffset for this, and
time is measured as:
(Itime-ItimeI+GLToffset)
Simply done to have one less hurdle for getting rid of itime_tr0.
(In fact, it looks like the old formula would be wrong for normal model
start ups where the user did not set itime_tr0, as it would default
to ItimeI. Hence the new one (where GLToffset defaults to 0) changes
results for GLT tracer/diagnostics in normal templates.)