remove parameter method emissions sect/reg scaling
...to be replaced shortly by gridded timestream file method.

This ended up removing so much code (e.g. tracers/
EmissionRegion.F90 and tracers/EmissionScenario.F90) that I
feel bad about Tom Clune's group's work on this. I even
considered keeping it commented out. But in the end I think
we should move on and the upcomming method will be easier for
both users and developers. If I was wrong to do this, and
there are folks using modelE at GSFC where they create tons
of sectors and regions, I could assist in putting this code
back in.