modified:   templates/E6TlernerF40.R
Changed "13824" to "13104" in itime_tr0.  As is, SF6_c tracer (when running
both SF6 and Lerner tracers turned on) is not initialized correctly.  SF6 and
SF6_c should be identical after a one month long simulation (nonlinearities in
annual mean emissions of SF6 do not kick in yet) but are not, unless said
change is made.