make TRACERS_ACETONE reproducible
TRACERS_ACETONE had a reproducibility problem because
the ocean source/sink is calculated before SURFACE is
called so trcSurfByVol is not yet calculated. This
commit fixes it by saving to rsfs (for now) the array
trcSurfByVol when TRACERS_ACETONE is defined. Potentially,
a future commit might only save this for tracers needed.

Also changing use of atmocn%wsavg to atmsrf%wsavg in that
ocean source/sink calculation

megan.F90 was just the removal of a stray comment character.