Bugfix affecting prescribed-dust perpetual-year runs.   The bug caused such runs to see transient (yearly varying) dust, if the input data contained multiple years.
The code for yearly varying dust was introduced as an option in
early 2017, at which time I failed to notice that KYEARD was being
zeroed out not only as part of its declaration, but also
after the XXX_yr logic in init_rad(), unfortunately.

Thanks to Kostas for noticing an effect of this bug.