W.I.P. replacing Acetone photolysis when tracer
Only change in results should be for TRACERS_ACETONE compiled.

Initial attempt to add two photolysis reactions to fastj, related
to Acetone tracer (meaning used when TRACERS_ACETONE is defined)
and to skip over the special treatment for acetone in the chemistry
when it is not a tracer. (Eventually, the hope is to delete that
treatment and hence the code will become cleaner with regards to
acetone -- e.g. the ifdefs I am introducing here should be
temporary). This involved allowing pressure-dependent cross-sections
to be read in and have fastj interpolates them exactly like the
temperature-dependent ones. It looked to me like that's how
fastj-X 7.? does it. But this is not a full implementation of
Fastj-X.

For now, these two reactions should be considered place-holders
until Drew S. and Kostas T. approve, even if nothing bad happens
with a simulation run like this.

Some tangentially-related cleanup:
(1) remove some older method reading of pressure dependencies,
    which did not seem to me to be doing anything as-is.
(2) remove code relating to option of "AR5_FASTJ_XSECS", which
    was intended to be a temporary method of reverting to older
    fastj settings from around 2014.