This is a quick hack that will allow the proper use of cpp instructions everywhere in the template text.
Current default behaviour is to treat statements between
Preprocessor Options
and
End Preprocessor Options
as a special text which is literally transferred to the rundeck.

With the new functionality, which is enabled by putting a line

  ###cpp_protected

at the beginning of the template, the entire text of the templete
is treated in the same way, i.e. the standard cpp preprocessing
is applied to all of it. The cpp instructions which need to be transferred
literally to the rundeck need to be escaped by replacing
 # with _ .