Changed the default values of some parameters to correspond to what is now used for the vast majority of runs.
This allows for shorter rundecks and templates.

No corresponding changes were made to any existing template because a new
set of templates is in the process of being constructed.    Those using
the current templates, feel free to edit them.

No effect on results, but be aware that text-to-float conversion
roundoff will cause any rundeck that explicitly specifies the current
default FT8OPX(dust)=1.3 will have roundoff differences compared to
one that relies upon the 1.3 in the code.