coupled_chem=0 I/O work
The units in the Koch code that reads chemistry fields when NOT
coupled to online chemistry were simplified and documented (using
my best guess). This amounted to removing the powers of 10
conversions online, in favor of converting units when making the
input file.

Thus this code update and the templates update (which swaps to new
files) must be done in tandem. The new input files for OFFLINE_CHEM
and OFFLINE_AERO were constructed from the same transient ensemble
currently in use for NINT aerosol/o3 input. Two caveats for now:
(1) both the yearly and climatological templates use climatology
files for now - though yearly could be made. (2) the C90 one is just
a placeholder (non-existent file).

The script used to convert OMA taijl output to coupled_chem=0 input
is also pushed with this commit [aux/OMA_out_to_coupleChem0_in.ksh]

Expecting no change in results for regression tested decks, but with
coupled_chem=0, definitely. In fact, the coupled_chem=0 case must
still be tested, but on the bright side, this might be the conclusion
of the coupled_chem=0 revitalization effort.