Corrected the diagnostic for the implied convergence of heat from ice and snow advection for the prescribed-ice (kocean==0) case.   It seems this diagnostic was not fully updated when the snow advection was separated from ice advection.
Though there should have been an indirect impact on qflux runs, there was not,
because an oversight elsewhere caused the calculation of the implied qfluxes to
omit this term.

The qflux calculation has been re-formulated and is about to be
committed to another branch.   It will be merged back after testing
is complete.