Allow 2nd NINT O3 input O3file2 (T. Cowan work)
Tyler Cowan's work to allow for a second "O3file" called
O3file2. If present, NINT ozone input is used from the surface to
the top of file O3file2 (as dictated by its ple meta data). Above
that level, the O3file's ozone is used up to its top. This could
be expanded in the future, but for now we require that the top
pressure edge in the O3file2 matches a pressure edge in the O3file,
for trivial merging of the two.

For example, this allows for quicker/easier replacement of
NINT O3 input from the output of a chemistry run: Its O3 can be
provided in O3file2 without dealing with offline merging with O3
above the GCM top, where we have no chemistry output.

This commit does not change the template input files, so no
change in results expected, as the code is backwards-compatible.