add 2 scripts to transate OMA output-->NINT input
The goal was a shell script anyone could run (with few dependencies;
uses NCO and ksh only) that would create from a year of OMA output
iall the needed NINT constituent input. One caveat is that the ozone
made with the script is only for tracersE3 and later models (requiring
O3file2 functionality). But the file gets tagged to remind you of that.
That's why I am pushing to tracersE3. Aerosols are OK for older models.

Executing aux/oneYearTracersToNINT.ksh without arguments will give
usage instructions, and an example script to show how to run the main
script e.g. in a loop over years is provided with a fair amount of comments:
aux/runTracersToNINT_example.ksh

I recommend running on a compute node with sufficient time requested.
Let me know of problems.